deta-backend/requirements.txt
fastapi
uvicorn
python-multipart
pdf2docx
aiofiles
python-docx
